fix(gateway): isolate transfer HTTP contract
